Thu, May 18, 2023, 2:33 PM +03:00

+++ finance/app.py
@@ -64,7 +64,7 @@ def buy():
         stocks = stock["name"]
         price = stock["price"]
 
-        total = int(price * shares)
+        total = float(price * shares)
 
         if shares < 0:
             return apology("Shares Error")
